+++ src/StringHelper.php
@@ -476,17 +476,15 @@ final class StringHelper
 
     /**
      * Split a string to array with non-empty lines.
-     * Whitespace from the beginning and end of a each line will be stripped by using {@see trim()}.
+     * Whitespace from the beginning and end of a each line will be stripped.
      *
      * @param string $string The input string.
      *
      * @return array
      */
-    public static function extractLines(string $string): array
+    public static function split(string $string): array
     {
-        $lines = preg_split('~(\r\n|\n\r|\n|\r)~', $string, -1, PREG_SPLIT_NO_EMPTY);
-        $lines = array_map('trim', $lines);
-        $lines = array_filter($lines, fn ($line) => $line !== '');
-        return array_values($lines);
+        $string = preg_replace('(^\s*|\s*$)', '', $string);
+        return preg_split('~\s*\R\s*~', $string, -1, PREG_SPLIT_NO_EMPTY);
     }
 }